RecipeBackend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-genai
instaloader
httpx
python-dotenv
pyktok
